% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ava.site.min.R
\name{ava.site.min}
\alias{ava.site.min}
\title{Minimum avalanching site}
\usage{
ava.site.min(x, y, height)
}
\arguments{
\item{x}{x parameter of the selected point}

\item{y}{y parameter of the selected point}

\item{height}{topography parameter}
}
\value{
the coordinates of the avalanching site.
            If there are more than one points satisfying avalanching requirement,
            return the one with minimum angle to the selected point
            If no avalanching site is found, return \code{c(-1,-1)}
}
\description{
Find out if there exists a avalanching site nearby. If not, return (-1,-1). If
it does, return the minimum avalanching site coordinates.
}
\details{
Avalanching is defined as the region that is equal to or higher than
            30 degrees to the highest point nearby.
}
\examples{
height=topo(100,100); ava.site.min(10,35,height)
}

