% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.sd.R
\name{is.sd}
\alias{is.sd}
\title{Shadow zone}
\usage{
is.sd(x, y, height)
}
\arguments{
\item{x}{x parameter of the selected point}

\item{y}{y parameter of the selected point}

\item{height}{topography parameter}
}
\value{
T if is in shadow zone; F if not
}
\description{
Check if a selected point is in shadow zone
}
\details{
Shadow zone is defined as the region that is within 15 degree
            to the highest point nearby.
}
\examples{
height=topo(100,100); is.sd(10,35,height)
}

