% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iteration.R
\name{iteration}
\alias{iteration}
\title{Run interations of sand transportation process}
\usage{
iteration(j, height)
}
\arguments{
\item{j}{Number of iterations}

\item{height}{topography parameter}
}
\value{
\code{height} after j iterations
}
\description{
Runs the whole program for j iterations on the topography parameter "height"
}
\details{
The program is conducted by the following steps:
            1. Generate topography paramter
            2. Select a random site
            3. Check if the site is in the shadow zone. If yes, go to next iteration
               if not, (and height is not 0), go to next steps
            4. Pick up one sand slab and deposit it down wind (along x axis) for 1 grid.
            5. Check the erosion site and deposition site if avalanch happens
               If avalanch requirements are satisfied, avalanch.
            6. Go to next iteration if j requires.
            Sand surface depositional probability is 0.6, and the probability for non-sand
            surface is 0.4
}
\examples{
height=topo(100,100)
iteration(1000,height)
}

